# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HaqLens Legal API",
    "version": "0.1.0",
    "description": "Legal-information preparation API. It does not provide legal advice. Evidence requests and responses are non-cacheable."
  },
  "paths": {
    "/health": { "get": { "responses": { "200": { "description": "Service is available" } } } },
    "/v1/extractions/facts": {
      "post": {
        "description": "Extract evidence-linked factual candidates. 64 KiB JSON limit; no-store response.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["evidence"],
                "properties": { "evidence": { "type": "array", "minItems": 1, "maxItems": 20 } },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Facts with fixture or Gemini provenance" },
          "400": { "description": "Invalid JSON" },
          "413": { "description": "Payload too large" },
          "415": { "description": "JSON required" },
          "422": { "description": "Invalid request" },
          "429": { "description": "Rate limited; Retry-After is set" },
          "503": { "description": "Rule-only safe mode" }
        }
      }
    },
    "/v1/routes/prepare": {
      "post": {
        "description": "Create a deterministic preparation result from reviewed evidence-linked facts. 64 KiB JSON limit; no-store response.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["evidence", "facts"],
                "properties": {
                  "evidence": { "type": "array", "minItems": 1, "maxItems": 100 },
                  "facts": { "type": "array", "minItems": 1, "maxItems": 100 }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Deterministic preparation result" },
          "400": { "description": "Invalid JSON" },
          "413": { "description": "Payload too large" },
          "415": { "description": "JSON required" },
          "422": { "description": "Invalid request" },
          "429": { "description": "Rate limited; Retry-After is set" },
          "500": { "description": "Safe internal error envelope" }
        }
      }
    },
    "/v1/briefs/document": {
      "post": {
        "description": "Return a bounded fixture brief whose citations refer only to submitted, source-exact text segments. 64 KiB JSON limit; no-store response. This endpoint does not determine legal rights, law, routes, or deadlines.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SegmentedDocument" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Source-linked document brief",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DocumentBrief" }
              }
            }
          },
          "400": { "description": "Invalid JSON" },
          "413": { "description": "Payload too large" },
          "415": { "description": "JSON required" },
          "422": { "description": "Invalid source-exact segmented document" },
          "429": { "description": "Rate limited; Retry-After is set" }
        }
      }
    },
    "/v1/comparisons/document": {
      "post": {
        "description": "Align two bounded segmented documents deterministically. Returns only source-scoped additions, removals, and changes; it does not provide legal interpretation. 64 KiB JSON limit; no-store response.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DocumentComparisonInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Deterministic source-aligned comparison",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DocumentComparison" }
              }
            }
          },
          "400": { "description": "Invalid JSON" },
          "413": { "description": "Payload too large" },
          "415": { "description": "JSON required" },
          "422": { "description": "Invalid source-exact segmented documents" },
          "429": { "description": "Rate limited; Retry-After is set" }
        }
      }
    },
    "/v1/dates/calculate": {
      "post": {
        "description": "Add an explicit calendar-day offset to a user-confirmed, source-linked ISO date. It does not infer legal deadlines, statutory periods, or consequences. 64 KiB JSON limit; no-store response.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DateCalculationInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Confirmed date or human-review result",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DateCalculationResult" }
              }
            }
          },
          "400": { "description": "Invalid JSON" },
          "413": { "description": "Payload too large" },
          "415": { "description": "JSON required" },
          "422": { "description": "Invalid calendar date or request" },
          "429": { "description": "Rate limited; Retry-After is set" }
        }
      }
    },
    "/v1/questions/document": {
      "post": {
        "description": "Return only a cited, source-exact document excerpt selected by deterministic lexical matching. Instruction-like or unsupported questions return needs_human_review; this endpoint does not provide legal advice, rights, deadlines, or consequences. 64 KiB JSON limit; no-store response.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/GroundedQuestionInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cited extractive answer or human-review result",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GroundedAnswer" }
              }
            }
          },
          "400": { "description": "Invalid JSON" },
          "413": { "description": "Payload too large" },
          "415": { "description": "JSON required" },
          "422": { "description": "Invalid source-exact segmented document or question" },
          "429": { "description": "Rate limited; Retry-After is set" }
        }
      }
    },
    "/v1/sources/{topic}": {
      "get": {
        "description": "Resolve a generic validated topic against the reviewed local official-source manifest. The service does not send document content to web search.",
        "parameters": [
          {
            "name": "topic",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["consumer_dispute", "employment_dispute", "legal_aid", "public_grievance"]
            }
          }
        ],
        "responses": {
          "200": { "description": "Reviewed local official sources" },
          "422": { "description": "Invalid topic" },
          "429": { "description": "Rate limited; Retry-After is set" }
        }
      }
    },
    "/v1/transcriptions": {
      "post": {
        "description": "Transcribe an accepted PDF/image through the opt-in provider. The response is always unconfirmed and must be reviewed by the user before analysis. No document is persisted.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TranscriptionRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Unconfirmed page transcription",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TranscriptionReview" }
              }
            }
          },
          "400": { "description": "Invalid JSON" },
          "413": { "description": "Payload too large" },
          "415": { "description": "JSON required" },
          "422": { "description": "Invalid transcription request" },
          "429": { "description": "Rate limited; Retry-After is set" },
          "503": { "description": "Transcription unavailable or provider failure" }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "DocumentSegment": {
        "type": "object",
        "required": ["id", "page", "sourceStart", "sourceEnd", "text"],
        "properties": {
          "id": { "type": "string", "pattern": "^segment-[1-9]\\d*$", "maxLength": 128 },
          "page": { "type": ["integer", "null"], "minimum": 1 },
          "sourceStart": { "type": "integer", "minimum": 0, "maximum": 59999 },
          "sourceEnd": { "type": "integer", "minimum": 1, "maximum": 60000 },
          "text": { "type": "string", "minLength": 1, "maxLength": 2000 }
        },
        "additionalProperties": false,
        "description": "Segment text must exactly equal text.slice(sourceStart, sourceEnd)."
      },
      "SegmentedDocument": {
        "type": "object",
        "required": ["sourceLabel", "text", "segments"],
        "properties": {
          "sourceLabel": { "type": "string", "minLength": 1, "maxLength": 200 },
          "text": { "type": "string", "minLength": 1, "maxLength": 60000 },
          "segments": {
            "type": "array",
            "minItems": 1,
            "maxItems": 250,
            "items": { "$ref": "#/components/schemas/DocumentSegment" }
          }
        },
        "additionalProperties": false
      },
      "DocumentCitation": {
        "type": "object",
        "required": ["segmentIds"],
        "properties": {
          "segmentIds": {
            "type": "array",
            "minItems": 1,
            "maxItems": 10,
            "uniqueItems": true,
            "items": { "type": "string", "pattern": "^segment-[1-9]\\d*$", "maxLength": 128 }
          }
        },
        "additionalProperties": false
      },
      "BriefItem": {
        "type": "object",
        "required": ["citation", "kind", "severity", "text"],
        "properties": {
          "citation": { "$ref": "#/components/schemas/DocumentCitation" },
          "kind": {
            "type": "string",
            "enum": ["summary", "risk", "uncertainty", "professional_question"]
          },
          "severity": { "type": ["string", "null"], "enum": ["low", "medium", "high", null] },
          "text": { "type": "string", "minLength": 1, "maxLength": 1000 }
        },
        "additionalProperties": false
      },
      "DocumentBrief": {
        "type": "object",
        "required": ["document", "items"],
        "properties": {
          "document": { "$ref": "#/components/schemas/SegmentedDocument" },
          "items": {
            "type": "array",
            "minItems": 1,
            "maxItems": 50,
            "items": { "$ref": "#/components/schemas/BriefItem" }
          }
        },
        "additionalProperties": false,
        "description": "Every citation segment ID must exist in document.segments."
      },
      "DocumentComparisonInput": {
        "type": "object",
        "required": ["left", "right"],
        "properties": {
          "left": { "$ref": "#/components/schemas/SegmentedDocument" },
          "right": { "$ref": "#/components/schemas/SegmentedDocument" }
        },
        "additionalProperties": false
      },
      "DocumentComparison": {
        "allOf": [
          { "$ref": "#/components/schemas/DocumentComparisonInput" },
          {
            "type": "object",
            "required": ["changes"],
            "properties": {
              "changes": {
                "type": "array",
                "maxItems": 500,
                "items": {
                  "type": "object",
                  "required": ["kind", "leftSegmentIds", "rightSegmentIds"],
                  "properties": {
                    "kind": { "type": "string", "enum": ["added", "removed", "changed"] },
                    "leftSegmentIds": { "type": ["array", "null"], "maxItems": 10 },
                    "rightSegmentIds": { "type": ["array", "null"], "maxItems": 10 }
                  },
                  "additionalProperties": false
                }
              }
            }
          }
        ],
        "description": "Each change is added, removed, or changed and references only the correct document side."
      },
      "DateCalculationInput": {
        "type": "object",
        "required": ["anchor", "offsetDays"],
        "properties": {
          "anchor": {
            "type": "object",
            "required": ["confirmed", "date", "citation"],
            "properties": {
              "confirmed": { "type": "boolean" },
              "date": { "type": "string", "format": "date" },
              "citation": { "$ref": "#/components/schemas/DocumentCitation" }
            },
            "additionalProperties": false
          },
          "offsetDays": { "type": "integer", "minimum": 0, "maximum": 3650 }
        },
        "additionalProperties": false
      },
      "DateCalculationResult": {
        "allOf": [
          { "$ref": "#/components/schemas/DateCalculationInput" },
          {
            "type": "object",
            "required": ["date", "status"],
            "properties": {
              "date": { "type": ["string", "null"], "format": "date" },
              "status": { "type": "string", "enum": ["confirmed", "needs_human_review"] }
            }
          }
        ],
        "description": "date is null unless the source anchor was explicitly confirmed."
      },
      "GroundedQuestionInput": {
        "type": "object",
        "required": ["document", "question"],
        "properties": {
          "document": { "$ref": "#/components/schemas/SegmentedDocument" },
          "question": { "type": "string", "minLength": 1, "maxLength": 1000 }
        },
        "additionalProperties": false
      },
      "GroundedAnswer": {
        "type": "object",
        "required": ["answer", "citation", "status"],
        "properties": {
          "answer": { "type": ["string", "null"], "minLength": 1, "maxLength": 2000 },
          "citation": {
            "oneOf": [{ "$ref": "#/components/schemas/DocumentCitation" }, { "type": "null" }]
          },
          "status": { "type": "string", "enum": ["answered", "needs_human_review"] }
        },
        "additionalProperties": false,
        "description": "answered requires a non-null answer and citation; needs_human_review requires both to be null. A returned citation always references the submitted document."
      },
      "TranscriptionRequest": {
        "type": "object",
        "required": ["dataBase64", "mimeType", "sourceLabel"],
        "properties": {
          "dataBase64": {
            "type": "string",
            "minLength": 4,
            "maxLength": 14000000,
            "pattern": "^[A-Za-z0-9+/]+={0,2}$"
          },
          "mimeType": {
            "type": "string",
            "enum": ["application/pdf", "image/jpeg", "image/png", "image/webp"]
          },
          "sourceLabel": { "type": "string", "minLength": 1, "maxLength": 200 }
        },
        "additionalProperties": false
      },
      "TranscriptionReview": {
        "type": "object",
        "required": ["confirmed", "pages", "sourceLabel"],
        "properties": {
          "confirmed": { "type": "boolean" },
          "sourceLabel": { "type": "string", "minLength": 1, "maxLength": 200 },
          "pages": {
            "type": "array",
            "minItems": 1,
            "maxItems": 250,
            "items": {
              "type": "object",
              "required": ["confidence", "page", "text"],
              "properties": {
                "confidence": { "type": "number", "minimum": 0, "maximum": 1 },
                "page": { "type": "integer", "minimum": 1 },
                "text": { "type": "string", "minLength": 1, "maxLength": 60000 }
              },
              "additionalProperties": false
            }
          }
        },
        "additionalProperties": false,
        "description": "confirmed is always false for provider output and must be set only after user review."
      }
    }
  }
}